requirements.txt
Django==1.11.3
image==1.5.9
olefile==0.44
Pillow==4.2.1
pytesseract==0.1.7
pytz==2017.2
